+++ tools/op-rights-demo.js
@@ -1,17 +1,6 @@
 const fs = require('fs');
 const path = require('path');
-
-function parseUserState(filePath) {
-  const p = filePath || path.join(__dirname, '..', 'app', 'user_state.yaml');
-  if (!fs.existsSync(p)) return null;
-  const data = fs.readFileSync(p, 'utf8').split(/\r?\n/);
-  let op = null;
-  data.forEach(line => {
-    const m = line.match(/^\s*op_level:\s*(.*)$/);
-    if (m) op = m[1].replace(/['"]/g, '');
-  });
-  return op;
-}
+const { parseUserState } = require('./user-state.js');
 
 function loadPermissions(filePath) {
   const p = filePath || path.join(__dirname, '..', 'permissions', 'op-permissions-expanded.json');
@@ -33,7 +22,8 @@ function showRights(level) {
 
 if (require.main === module) {
   const argLevel = process.argv[2];
-  const level = argLevel || parseUserState() || 'OP-0';
+  const state = parseUserState();
+  const level = argLevel || (state ? state.op_level : null) || 'OP-0';
   showRights(level);
 }
